xforms module: Fixed bug where serialization was not going beyond 127 number of users.

+++ src/org/openmrs/module/xforms/serialization/DefaultUserSerializer.java
@@ -50,13 +50,13 @@ public class DefaultUserSerializer {
 			DataOutputStream dos = new DataOutputStream(os);
 			
 			if(data == null){
-        		dos.writeByte(0);
+        		dos.writeInt(0);
         		return;
         	}
 			
 			List<XformUser> users = (List<XformUser>)data; //This will always be a list of XFormUser
 
-			dos.writeByte(users.size());
+			dos.writeInt(users.size());
 			for(XformUser user : users)
 				serialize(user,dos);
